linux_inet_diag: fix Wshorten-64-to-32 warnings
POSIX and glibc 2.2+ declare the hostlen and servlen
args of getnameinfo(3) to be socklen_t, not size_t,
so favor socklen_t for those calculations.

While we're at it, nlmsg_len is u32, too, so cast it
as such to avoid the warning.

Tested on clang version 3.5-1ubuntu1 on x86-64