doc: update URLs and references
Neither the GCC nor libatomic_ops URLs are valid, anymore.
Update them to the latest versions and ensure the URLs themselves
are visible in documentation as links to external sites.

Additionally, the current cgit installation on bogomips.org will
soon be replaced by a CSS-free web-based viewer.

Also, correct the reference to "LGPL-2.1+" while we're at it
and add references to the nntp:// mail archives.